Change liveness and readiness probes for the ovn-northd POD
Before this patch liveness and readiness probes for the ovn-northd POD
were just checking PID of the ovn-northd process.

This patch changes it to use custom script which checks ovn-northd
status with `ovs-appctl status` command.
According to the OVN documentation [1] this command:

  Prints this server’s status. Status will be "active" if ovn-northd has
  acquired OVSDB lock on SB DB, "standby" if it has not or "paused" if
  this instance is paused.

As `ovs-appctl status` command returns `0` as exit code even if status
of the ovn-northd is e.g. "standby", this patch adds custom script which
is used by the liveness and readiness probes and which returns `1` in
case when status is anything else then `active`.

[1] https://man7.org/linux/man-pages/man8/ovn-northd.8.html

Closes: #OSPRH-8042
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>